Complete template strings exercise and Traceur processor.
Traceur processor now correctly transpiles both the submission and the solution.

Turns out date/time is not a great idea since it can vary by a millisecond or so between running the solution and running the submission. So switched to lowercasing instead.

Now using a temporary directory instead of .transpiled.js, since otherwise it will display the transpiled version alongside the ES6 solution, which is bad.